move zip implementation to wdl_packager
